[fix] 기록의 result도 없는 경우에 초기화
 기록의 result도 없는 경우에 초기화

+++ server.js
@@ -65,7 +65,7 @@ cron.schedule('0 1 0 * * *', async () => {
       date.setDate(date.getDate() - 1);
       const record = await Record.findByPk(`${String(date.getFullYear())}-${String(date.getMonth() + 1).padStart(2, '0')}-${String(date.getDate()).padStart(2, '0')}`);
       if (record) {
-        if (Object.keys(record.metadata.log).length === 0) {
+        if ((Object.keys(record.result).length === 0) && (Object.keys(record.metadata.log).length === 0)) {
           await record.destroy({ transaction: t });
         } else {
           await utils.setResult(t, record.date, record.metadata.log);
